Исправлены замечания ревьюера: - убран закомментированный код; - для цены товара установлоен тип BigDecimal вместо float

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/dto/shopping/store/ProductDto.java
@@ -1,8 +1,9 @@
 package ru.yandex.practicum.commerce.dto.shopping.store;
 
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record ProductDto (UUID productId, String productName, String description, String imageSrc,
-                          QuantityState quantityState, ProductState productState, ProductCategory productCategory, float price) {
+                          QuantityState quantityState, ProductState productState, ProductCategory productCategory, BigDecimal price) {
 }
